+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.2.0] - 19/02/2018
+
+* `MaterialSearchInput` flutter upgrade [#8](https://github.com/ianldgs/material_search/issues/8)
+* Example app [#7](https://github.com/ianldgs/material_search/issues/7)
+* Fix bug when ommiting `getResults` option [#6](https://github.com/ianldgs/material_search/issues/6)
+* `MaterialSearchInput` form validation [#4](https://github.com/ianldgs/material_search/issues/4)
+* Add a few unit test assertions [#1](https://github.com/ianldgs/material_search/issues/1)
+
 ## [0.1.0] - 22/01/2018
 
 `getResults` option